requirements.txt
django-redis==4.2.0
django-rest-swagger==0.3.2
django-secure==1.0.1
django-sslserver==0.15
djangorestframework==3.1.3
django-cors-headers==1.1.0